\name{DesignPattern-package}
\alias{DesignPattern-package}
\alias{DesignPattern}
\docType{package}
\title{
DESIGN PATTERNS IN R
}
\description{
Design patterns are building blocks of reusable object-oriented software
projects. This package provides data structures and tools to implement,
study and reuse design patterns in R, especially the ones introduced in
the book 'Gang of Four'. 
}
\details{
\tabular{ll}{
Package: \tab DesignPattern\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2009-10-23\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}
\author{
Jitao David Zhang <j.zhang@dkfz.de>, supported by Stefan Wiemann and
Wolfgang Huber

Maintainer: Jitao David Zhang <j.zhang@dkfz.de>
}
\references{
Design Patterns: Elements of Reusable Object-Oriented Software. Erich
Gamma, Richard Helm, Ralph Johnson and John Vlissides. Addison-Wesley
Publishing Company, 1995. (well-known as 'Gang of Four' or 'GoF')
}
\keyword{ package }
%\examples{
%}
